Moves CLI to [[bin]], adds [lib] target to sage-cli crate
