+++ crates/sage-cli/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod input;
+pub mod runner;
+pub mod telemetry;
